% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classData}
\alias{classData}
\title{A dataset for classification}
\format{
A list containing a protein expression matrix and a response vector
\describe{
\item{Exp}{a protein expression matrix}
\item{Label}{a response vector}
}
}
\usage{
classData
}
\description{
A dataset for classification
}
\keyword{datasets}
