% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PCGroups}
\alias{PCGroups}
\title{Protein complexes.}
\format{
A data frame with 3512 rows and 6 variables:
\describe{
\item{ComplexID}{ID of the protein complex}
\item{ComplexName}{name of the protein complex}
\item{Organism}{organism}
\item{UniprotID}{Uniprot IDs of the proteins in the protein complex}
\item{EntrezID}{Entrez IDs of the proteins in the protein complex}
\item{GeneSymbol}{gene symbols of the proteins in the protein complex}
}
}
\source{
\url{https://mips.helmholtz-muenchen.de/corum/}
}
\usage{
PCGroups
}
\description{
A dataset containing the protein complexes
}
\keyword{datasets}
