% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCLassoReg-package.R
\docType{package}
\name{PCLassoReg}
\alias{PCLassoReg}
\title{\packageTitle{PCLassoReg}}
\description{
\packageDescription{PCLassoReg}
}
\details{
\packageIndices{PCLasso} The PCLasso model accepts a protein expression
matrix, survival data, and protein complexes for training the prognostic
model, and makes predictions for new samples and identifies risk protein
complexes associated with survival.

The PCLasso2 model accepts a protein expression matrix, a response vector,
and protein complexes for training the classification model, and makes
predictions for new samples and identifies risk protein complexes associated
with classes.

Both PCLasso and PCLasso2 use \code{grLasso} as the penalty function. The
other two penalties \code{grSCAD} and \code{grMCP} can also be used for model
construction and risk protein complex identification. The package also
provides methods for plotting coefficient paths and cross-validation curves.
}
\references{
PCLasso2: a protein complex-based, group Lasso-logistic model for risk
protein complex discovery. To be published.

PCLasso: a protein complex-based group lasso-Cox model for accurate prognosis
and risk protein complex discovery. Brief Bioinform, 2021.

Park, H., Niida, A., Miyano, S. and Imoto, S. (2015) Sparse overlapping group
lasso for integrative multi-omics analysis. Journal of computational biology:
a journal of computational molecular cell biology, 22, 73-84.
}
