% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{survivalData}
\alias{survivalData}
\title{A dataset for prognostic model}
\format{
A list containing a protein expression matrix and survival data
\describe{
\item{Exp}{a protein expression matrix}
\item{survData}{Survival data. The first column is a binary variable with
1 indicating that the event has occurred and 0 indicating(right); the
second column is the time on study (follow up time)}
}
}
\usage{
survivalData
}
\description{
A dataset for prognostic model
}
\keyword{datasets}
